Cambiado Integer por Long

+++ src/main/java/com/fot/backcontask/dao/GenericDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.NoRepositoryBean;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
 @NoRepositoryBean
-public interface GenericDAO<D> extends PagingAndSortingRepository<D, Integer> {
+public interface GenericDAO<D> extends PagingAndSortingRepository<D, Long> {
 
 }
